Fix issue where the daily condition is sometimes not calculated
The issue happens when the whole day is ATTR_CONDITION_PARTLYCLOUDY, but cloud_cover_avg is still below the threshold. This commit adds a simple "shortcut" right at the beginning so that if the whole day has the same condition for each hour, we just use this condition for the day, before even starting more sophisticated evaluations.